+++ core/src/test/scala/org/apache/spark/benchmark/BenchmarkBase.scala
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.spark.util
+package org.apache.spark.benchmark
 
 import java.io.{File, FileOutputStream, OutputStream}
